+++ CHANGELOG.md
@@ -4,6 +4,22 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+
+## [0.7.7]
+
+### Changed
+
+  * Refactor: New objectdir module now contains all objectdir related code
+  * Add validation for objectdirs:
+    * Check structure
+    * Check if object dir name matches IDs in TEI and LIDO Files
+    * Check if object dir name matches identifier in DC.xml
+  * Underscores now are valid in datastream IDs
+  * ObjectCSVManager now has a get_mainresource method 
+  * Code cleanup  
+  
+
 ## [0.7.6] - 2025-11-21
 
 ### Changed